cardvision/requirements.txt
ultralytics>=8.2.0
flask>=3.0.0
flask-cors>=4.0.0
opencv-python>=4.9.0
Pillow>=10.0.0
watchdog>=4.0.0
numpy>=1.24.0
PyYAML>=6.0
